Resolve the Clang-Tidy findings that were failing CI
The Clang-Tidy job has been red since the check was introduced, reporting 60
diagnostics across twelve files.

Most are widened arithmetic: multiplications evaluated in int or unsigned int
and only then assigned to size_t or qint64, and casts applied after the
arithmetic rather than before it. Those now widen an operand first, so the
result no longer depends on the narrower type not overflowing.

The remainder are narrowing conversions made explicit, reserve() before
push_back loops, and a value parameter passed as a const reference.

ReviewDialog::result and VideoReviewDialog::result shadowed QDialog::result, so
calling either through a QDialog pointer silently returned the int dialog code
instead. Both are renamed to reviewResult.

updateModelPathFromSelection had two identical branches; the placeholder case is
now the only special one.

Three catch blocks stay empty on purpose - an unreadable orientation tag, a
best-effort second frame count, and a probe that decides an input size. Each
says why and carries a targeted NOLINT. The cv::Mat parameters that look
copyable are also annotated: the header is refcounted and taking it by value is
what keeps the pixels shared and writable.

+++ src/ModelDownloader.cpp
@@ -160,7 +160,8 @@ namespace cloakframe
     bool ensureBuiltinModelAvailable(
         QWidget *parent, const BuiltinModel &model, const QString &destPath)
     {
-        const auto sizeMb = QString::number(model.approxBytes / 1024.0 / 1024.0, 'f', 1);
+        const auto sizeMb =
+            QString::number(static_cast<double>(model.approxBytes) / 1024.0 / 1024.0, 'f', 1);
         const auto answer = QMessageBox::question(parent,
             QCoreApplication::translate("cloakframe::MainWindow", "Download Model"),
             builtinModelConsentText(model).arg(model.fileName, sizeMb),
@@ -176,7 +177,8 @@ namespace cloakframe
     bool ensurePlateModelAvailable(QWidget *parent, const QString &destPath)
     {
         const auto &model = plateModel();
-        const auto sizeMb = QString::number(model.approxBytes / 1024.0 / 1024.0, 'f', 1);
+        const auto sizeMb =
+            QString::number(static_cast<double>(model.approxBytes) / 1024.0 / 1024.0, 'f', 1);
         const auto answer = QMessageBox::question(parent,
             QCoreApplication::translate("cloakframe::MainWindow", "Download Model"),
             QCoreApplication::translate("cloakframe::MainWindow",
@@ -233,7 +235,7 @@ namespace cloakframe
                 "Only load ONNX models from sources you trust.\n\nModel: %1\nSize: %2 "
                 "MB\n\nContinue?")
                 .arg(info.fileName())
-                .arg(QString::number(info.size() / 1024.0 / 1024.0, 'f', 1)),
+                .arg(QString::number(static_cast<double>(info.size()) / 1024.0 / 1024.0, 'f', 1)),
             QMessageBox::Yes | QMessageBox::No,
             QMessageBox::No);
         return answer == QMessageBox::Yes;
